Repaired Harpoon Missile upgrade
The harpoon missile functionality is believed to be fixed.  More tests are required.   The games no longer crashes when performing the remove ACTION condition.   The splash damage on critical and destroyed ship also no longer crashes the game.    More testing should be made to see if damage is allocated properly.